more refactorings

+++ core/data/src/main/java/com/drodobyte/core/data/model/Id.kt
@@ -1,4 +1,4 @@
 package com.drodobyte.core.data.model
 
 @JvmInline
-value class Id(val id: Int)
+value class Id(val id: Long)
